requirements.txt
# --- Your Core Tools ---
PyQt5
numpy>=1.24
tqdm>=4.66
typing-extensions>=4.9
pandas
ipython

# --- Parsing & Document Loading ---
pymupdf>=1.23
docling>=0.1.7
lark>=1.1.0

# --- AI & Machine Learning ---
torch>=2.1
sentence-transformers>=2.6
openai>=1.30.0

# --- LangChain Ecosystem ---
langchain>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.2.0
langchain-openai>=0.2.0
langchain-text-splitters

# --- Vector Database ---
chromadb>=0.4.0
langchain-chroma>=0.1.4

# --- Utilities ---
python-dotenv>=1.0.1
tiktoken>=0.7.0